MAINT: Start restructuring mathematics folder: collect tools for measurements.
